requirements.txt
flask
flask-cors
gunicorn
python-jose
